refactor: 지출 합계 반환 타입을 Int에서 Long으로 변경
지출 금액의 총합 계산 시 발생할 수 있는 데이터 오버플로우를 방지하기 위해 `ExpenseDao`의 `getTotalExpenseByDiary` 함수 반환 타입을 `Flow<Int>`에서 `Flow<Long>`으로 수정합니다.

**주요 변경 사항:**

*   **`ExpenseDao.kt`**:
    *   `getTotalExpenseByDiary` 함수의 반환 타입을 `Flow<Long>`으로 변경하여 보다 큰 범위의 금액 합계를 처리할 수 있도록 개선했습니다.

+++ app/src/main/java/com/picday/diary/data/diary/dao/ExpenseDao.kt
@@ -20,5 +20,5 @@ interface ExpenseDao {
     fun getExpensesByDiaryId(diaryId: String): Flow<List<ExpenseEntity>>
 
     @Query("SELECT COALESCE(SUM(amount), 0) FROM expense WHERE diaryId = :diaryId")
-    fun getTotalExpenseByDiary(diaryId: String): Flow<Int>
+    fun getTotalExpenseByDiary(diaryId: String): Flow<Long>
 }
